fixed vga

+++ src/port.c
@@ -4,7 +4,7 @@ void outb(uint16_t port, uint8_t value) {
     __asm__ volatile("outb %0, %1" : : "a"(value), "Nd"(port));
 }
 
-uint16_t inb(uint32_t port) {
+uint8_t inb(uint16_t port) {
     unsigned char result;
     __asm__("in %%dx, %%al" : "=a" (result) : "d" (port));
     return result;
